+++ tests/testthat/test-docx.R
@@ -2161,6 +2161,63 @@ test_that("docx-60: Three level stub and indentation work as expected.", {
   }
 })
 
+test_that("docx-61: Table with blank_before works as expected.", {
+  
+  if (dev) {
+  fp <- file.path(base_path, "docx/test61.docx")
+  
+  
+  # Setup
+  subjid <- 100:109
+  name <- c("Quintana, Gabriel", "Allison, Blas", "Minniear, Presley",
+            "al-Kazemi, Najwa", "Schaffer, Ashley", "Laner, Tahma",
+            "Perry, Sean", "Crews, Deshawn Joseph", "Person, Ladon here is some more",
+            "Smith, Shaileigh and \nmore and more and even more and more and more")
+  sex <- c("M", "F", "F", "M", "M", "F", "M", "F", "F", "M")
+  age <- c(41, 53, 43, 39, 47, 52, 21, 38, 62, 26)
+  arm <- c(rep("A", 5), rep("B", 5))
+  
+  # Create data frame
+  df <- data.frame(subjid, name, sex, age, arm)
+  
+  
+  tbl1 <- create_table(df, first_row_blank = FALSE, borders = "all") %>%
+    define(subjid, label = "Subject ID", align = "left", width = 5) %>%
+    define(name, label = "Subject Name", width = 1) %>%
+    define(sex, label = "Sex") %>%
+    define(age, label = "Age") %>%
+    define(arm, label = "Arm",
+           blank_before = TRUE,
+           dedupe = TRUE,
+           align = "right") #%>%
+  # spanning_header(sex, arm, label = "Here is a spanning header")
+  
+  
+  rpt <- create_report(fp, output_type = "DOCX", font = fnt, font_size = fsz) %>%
+    page_header(left = "Experis", right = c("Study ABC", "Status: Closed")) %>%
+    # options_fixed(line_count = 46) %>%
+    titles("Table 1.0", "Analysis Data Subject Listing\n And more stuff",
+           "Safety Population", align = "center", bold = TRUE) %>%
+    footnotes("Program Name: table1_0.R",
+              "Here is a big long footnote that is going to wrap\n at least once") %>%
+    page_footer(left = "Time", center = "Confidential",
+                right = "Page [pg] of [tpg]") %>%
+    add_content(tbl1)
+  
+  
+  res <- write_report(rpt)
+  res
+  expect_equal(file.exists(fp), TRUE)
+  
+  } else {
+    
+    expect_equal(TRUE, TRUE) 
+    
+  }
+  
+  
+})
+
 # User Tests --------------------------------------------------------------